move components out of app dir
