+++ client/apps/event/edit/index.jsx
@@ -4,15 +4,15 @@ import { Col, Row } from 'react-styled-flexboxgrid'
 import EditDate from './components/edit_date.js'
 import EventDate from '../show/components/date.jsx'
 import EditVenue from './components/edit_venue.js'
-import TextInput from '../../components/forms/text_input.js'
-import ValidationError from '../../components/forms/validation_error.js'
-import { EditImagesList } from '../../components/forms/images/edit_images_list.jsx'
-import { EditLinkList } from '../../components/forms/links/edit_link_list.jsx'
-import { EditNav } from '../../components/forms/edit_nav.jsx'
-import { FileInput } from '../../components/forms/file_input/index.jsx'
-import { PlainText } from '../../components/forms/rich_text/plain_text.jsx'
-import { RichText } from '../../components/forms/rich_text/index.jsx'
-import { Venue } from '../../components/venue/index.jsx'
+import TextInput from '../../../components/forms/text_input.js'
+import ValidationError from '../../../components/forms/validation_error.js'
+import { EditImagesList } from '../../../components/forms/images/edit_images_list.jsx'
+import { EditLinkList } from '../../../components/forms/links/edit_link_list.jsx'
+import { EditNav } from '../../../components/forms/edit_nav.jsx'
+import { FileInput } from '../../../components/forms/file_input/index.jsx'
+import { PlainText } from '../../../components/forms/rich_text/plain_text.jsx'
+import { RichText } from '../../../components/forms/rich_text/index.jsx'
+import { Venue } from '../../../components/venue/index.jsx'
 require('./index.scss')
 
 export class EventEdit extends Component {
