+++ client/components/forms/images/edit_images_list.jsx
