+++ client/components/forms/rich_text/index.scss
