+++ client/components/icons/icon_menu.js
