+++ client/components/items_list/index.scss
