+++ client/styles/_apps.scss
@@ -1,0 +1,1 @@
+@import "../apps/event/index";
