+++ client/styles/_components.scss
@@ -1,7 +1,3 @@
-// Apps
-@import "../apps/event/index";
-
-// Components
-@import "../apps/components/forms/file_input/index";
-@import "../apps/components/forms/images/index";
-@import "../apps/components/items_list/index";
+@import "../components/forms/file_input/index";
+@import "../components/forms/images/index";
+@import "../components/items_list/index";
